fix: bug

+++ src/base/s2o.hpp
@@ -146,7 +146,7 @@ namespace orm {
   }
   template <>
   bool s2o<tm>(const std::string& str, tm& value) {
-	short year = 0, month = 0, day = 0, hour = 0, min = 0, sec = 0;
+	int year = 0, month = 0, day = 0, hour = 0, min = 0, sec = 0;
 	if (sscanf(str.c_str(), RES_DATE_FORMAT, &year, &month, &day, &hour, &min, &sec) != 6) {
 	  return false;
 	}
